１．各状態の実装 Attack,Hurt,Run,Die ２．UITextを追加 ３．敵のAI作成中

+++ Assets/Scripts/Role/RoleInfo/RoleInfoBase.cs
@@ -10,7 +10,7 @@ public class RoleInfoBase
     /// <summary>
     /// キャラクター　サーバーID
     /// </summary>
-    public int RoleServerID;
+    public long RoleServerID;
 
     /// <summary>
     /// キャラクターID（ロケーションID）
